# packages/contracts/openapi/plan-service.json
{
  "components": {
    "schemas": {
...
      "CalendarCreate": {
        "properties": {
          "code": {
            "description": "Не меняется после создания: на код ссылается DEFAULT_CALENDAR",
            "examples": [
              "moscow-5day"
            ],
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Code",
            "type": "string"
          },
          "holidays": {
            "description": "Нерабочие праздничные дни; у каждой даты — источник",
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Holidays",
            "type": "array"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "default": "Europe/Moscow",
            "description": "IANA, например Europe/Moscow",
            "title": "Timezone",
            "type": "string"
          },
          "weekend_days": {
            "description": "Выходные по ISO: понедельник = 1, воскресенье = 7",
            "items": {
              "type": "integer"
            },
            "title": "Weekend Days",
            "type": "array"
          },
          "work_hours": {
            "$ref": "#/components/schemas/WorkHours"
          }
        },
        "required": [
          "code",
          "name",
          "weekend_days",
          "work_hours"
        ],
        "title": "CalendarCreate",
        "type": "object"
      },
      "CalendarRead": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "holidays": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Holidays",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weekend_days": {
            "items": {
              "type": "integer"
            },
            "title": "Weekend Days",
            "type": "array"
          },
          "work_hours": {
            "$ref": "#/components/schemas/WorkHours"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "timezone",
          "weekend_days",
          "holidays",
          "work_hours",
          "created_at",
          "updated_at"
        ],
        "title": "CalendarRead",
        "type": "object"
      },
      "CalendarUpdate": {
        "description": "Частичное изменение. Списки заменяются целиком: праздник убирается новым списком.",
        "properties": {
          "holidays": {
            "anyOf": [
              {
                "items": {
                  "format": "date",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holidays"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "weekend_days": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekend Days"
          },
          "work_hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkHours"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CalendarUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_CalendarRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CalendarRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Всего записей, удовлетворяющих фильтру",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CalendarRead]",
        "type": "object"
      },
...
      "Page_RuleRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RuleRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Всего записей, удовлетворяющих фильтру",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RuleRead]",
        "type": "object"
      },
...
      "RequiredGroup": {
        "description": "Группа «любой из»: выполнена, если суммарно единиц классов не меньше min.",
        "properties": {
          "any_of": {
            "examples": [
              [
                "roller",
                "bulldozer"
              ]
            ],
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Any Of",
            "type": "array"
          },
          "min": {
            "minimum": 1.0,
            "title": "Min",
            "type": "integer"
          }
        },
        "required": [
          "any_of",
          "min"
        ],
        "title": "RequiredGroup",
        "type": "object"
      },
...
      "RuleCreate": {
        "properties": {
          "allowed": {
            "description": "Не вызывает D3",
            "items": {
              "type": "string"
            },
            "title": "Allowed",
            "type": "array"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "min_sessions": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Сколько рабочих сессий подряд держится условие; по умолчанию 2",
            "title": "Min Sessions"
          },
          "required": {
            "items": {
              "$ref": "#/components/schemas/RequiredGroup"
            },
            "title": "Required",
            "type": "array"
          },
          "signature": {
            "$ref": "#/components/schemas/Signature"
          },
          "stage_id": {
            "description": "У этапа не больше одного правила",
            "format": "uuid",
            "title": "Stage Id",
            "type": "string"
          }
        },
        "required": [
          "stage_id"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
      "RuleRead": {
        "properties": {
          "allowed": {
            "items": {
              "type": "string"
            },
            "title": "Allowed",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "min_sessions": {
            "title": "Min Sessions",
            "type": "integer"
          },
          "required": {
            "items": {
              "$ref": "#/components/schemas/RequiredGroup"
            },
            "title": "Required",
            "type": "array"
          },
          "signature": {
            "$ref": "#/components/schemas/Signature"
          },
          "stage_id": {
            "format": "uuid",
            "title": "Stage Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "description": "Растёт при каждой правке; попадает в отклонение",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "stage_id",
          "version",
          "required",
          "allowed",
          "signature",
          "min_sessions",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "RuleRead",
        "type": "object"
      },
      "RuleUpdate": {
        "description": "Частичное изменение. Списки и сигнатура заменяются целиком.",
        "properties": {
          "allowed": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "min_sessions": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Sessions"
          },
          "required": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RequiredGroup"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required"
          },
          "signature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Signature"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RuleUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkHours": {
        "properties": {
          "end": {
            "examples": [
              "23:00"
            ],
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "End",
            "type": "string"
          },
          "start": {
            "examples": [
              "07:00"
            ],
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "WorkHours",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Сервис плана: объекты, справочник работ, календарный график, правила «этап → техника». Единственное место, где план можно изменить.",
    "title": "plan-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/plan/calendars": {
      "get": {
        "operationId": "list_calendars_api_v1_plan_calendars_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CalendarRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Рабочие календари",
        "tags": [
          "Календари"
        ]
      },
      "post": {
        "operationId": "create_calendar_api_v1_plan_calendars_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Создать календарь",
        "tags": [
          "Календари"
        ]
      }
    },
    "/api/v1/plan/calendars/{calendar_id}": {
      "get": {
        "operationId": "get_calendar_api_v1_plan_calendars__calendar_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "calendar_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Calendar Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Календарь",
        "tags": [
          "Календари"
        ]
      },
      "patch": {
        "description": "Выходные, праздники, рабочие часы или часовой пояс: у всех объектов на календаре растёт `plan_version`, и в analysis уходит сигнал «пересчитай».",
        "operationId": "update_calendar_api_v1_plan_calendars__calendar_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "calendar_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Calendar Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Изменить календарь",
        "tags": [
          "Календари"
        ]
      }
    },
...
    "/api/v1/plan/rules": {
      "get": {
        "operationId": "list_rules_api_v1_plan_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "stage_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage Id"
            }
          },
          {
            "in": "query",
            "name": "object_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RuleRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Правила этапов",
        "tags": [
          "Правила этапов"
        ]
      },
      "post": {
        "description": "Коды классов проверяются по `equipment_classes.yaml`. Версия правила и `plan_version` объекта растут, в analysis уходит сигнал «пересчитай». `X-Actor` — кто правит, в журнал.",
        "operationId": "create_rule_api_v1_plan_rules_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Создать правило этапа",
        "tags": [
          "Правила этапов"
        ]
      }
    },
    "/api/v1/plan/rules/{rule_id}": {
      "delete": {
        "description": "По этапу без правила не проверяются D1, D2, D8, D9. `plan_version` растёт.",
        "operationId": "delete_rule_api_v1_plan_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Удалить правило этапа",
        "tags": [
          "Правила этапов"
        ]
      },
      "get": {
        "operationId": "get_rule_api_v1_plan_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Правило этапа",
        "tags": [
          "Правила этапов"
        ]
      },
      "patch": {
        "description": "Коды классов проверяются по `equipment_classes.yaml`. Версия правила и `plan_version` объекта растут, в analysis уходит сигнал «пересчитай». `X-Actor` — кто правит, в журнал.",
        "operationId": "update_rule_api_v1_plan_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Изменить правило этапа",
        "tags": [
          "Правила этапов"
        ]
      }
    },
...
  }
}